+++ BUILD
@@ -41,7 +41,7 @@ genrule(
 
 java_library(
     name = "console",
-    srcs = glob(["*.java"]) + [":version"],
+    srcs = glob(["*.java", "command/*.java", "common/*.java"]) + [":version"],
     deps = [
         "@graknlabs_client_java//:client-java",
         "@graknlabs_graql//java:graql",
